+++ ahbottomnavigation/src/main/java/com/aurelhubert/ahbottomnavigation/notification/AHNotification.java
@@ -12,7 +12,7 @@ import java.util.List;
 /**
  * @author repitch
  */
-public class Notification implements Parcelable {
+public class AHNotification implements Parcelable {
 
     @Nullable
     private String text; // can be null, so notification will not be shown
@@ -23,11 +23,11 @@ public class Notification implements Parcelable {
     @ColorInt
     private int backgroundColor; // if 0 then use default value
 
-    private Notification() {
+    public AHNotification() {
         // empty
     }
 
-    private Notification(Parcel in) {
+    private AHNotification(Parcel in) {
         text = in.readString();
         textColor = in.readInt();
         backgroundColor = in.readInt();
@@ -49,18 +49,14 @@ public class Notification implements Parcelable {
         return backgroundColor;
     }
 
-    public static Notification justText(String text) {
-        return newBuilder().setText(text).build();
+    public static AHNotification justText(String text) {
+        return new Builder().setText(text).build();
     }
 
-    public static Builder newBuilder() {
-        return new Notification().new Builder();
-    }
-
-    public static List<Notification> generateEmptyList(int size) {
-        List<Notification> notificationList = new ArrayList<>();
+    public static List<AHNotification> generateEmptyList(int size) {
+        List<AHNotification> notificationList = new ArrayList<>();
         for (int i = 0; i < size; i++) {
-            notificationList.add(Notification.newBuilder().build());
+            notificationList.add(new AHNotification());
         }
         return notificationList;
     }
@@ -77,40 +73,47 @@ public class Notification implements Parcelable {
         dest.writeInt(backgroundColor);
     }
 
-    public class Builder {
-        private Builder() {
-            // empty
-        }
+    public static class Builder {
+        @Nullable
+        private String text;
+        @ColorInt
+        private int textColor;
+        @ColorInt
+        private int backgroundColor;
 
         public Builder setText(String text) {
-            Notification.this.text = text;
+            this.text = text;
             return this;
         }
 
         public Builder setTextColor(@ColorInt int textColor) {
-            Notification.this.textColor = textColor;
+            this.textColor = textColor;
             return this;
         }
 
         public Builder setBackgroundColor(@ColorInt int backgroundColor) {
-            Notification.this.backgroundColor = backgroundColor;
+            this.backgroundColor = backgroundColor;
             return this;
         }
 
-        public Notification build() {
-            return Notification.this;
+        public AHNotification build() {
+            AHNotification notification = new AHNotification();
+            notification.text = text;
+            notification.textColor = textColor;
+            notification.backgroundColor = backgroundColor;
+            return notification;
         }
     }
 
-    public static final Creator<Notification> CREATOR = new Creator<Notification>() {
+    public static final Creator<AHNotification> CREATOR = new Creator<AHNotification>() {
         @Override
-        public Notification createFromParcel(Parcel in) {
-            return new Notification(in);
+        public AHNotification createFromParcel(Parcel in) {
+            return new AHNotification(in);
         }
 
         @Override
-        public Notification[] newArray(int size) {
-            return new Notification[size];
+        public AHNotification[] newArray(int size) {
+            return new AHNotification[size];
         }
     };