+++ demo/src/main/java/com/aurelhubert/ahbottomnavigation/demo/DemoActivity.java
@@ -1,6 +1,9 @@
 package com.aurelhubert.ahbottomnavigation.demo;
 
 import android.animation.Animator;
+import android.content.Context;
+import android.content.Intent;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.support.design.widget.FloatingActionButton;
@@ -8,7 +11,7 @@ import android.support.design.widget.Snackbar;
 import android.support.v4.content.ContextCompat;
 import android.support.v4.view.animation.LinearOutSlowInInterpolator;
 import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
+import android.support.v7.app.AppCompatDelegate;
 import android.view.View;
 import android.view.animation.OvershootInterpolator;
 
@@ -16,7 +19,7 @@ import com.aurelhubert.ahbottomnavigation.AHBottomNavigation;
 import com.aurelhubert.ahbottomnavigation.AHBottomNavigationAdapter;
 import com.aurelhubert.ahbottomnavigation.AHBottomNavigationItem;
 import com.aurelhubert.ahbottomnavigation.AHBottomNavigationViewPager;
-import com.aurelhubert.ahbottomnavigation.notification.Notification;
+import com.aurelhubert.ahbottomnavigation.notification.AHNotification;
 
 import java.util.ArrayList;
 
@@ -28,6 +31,7 @@ public class DemoActivity extends AppCompatActivity {
 	private ArrayList<AHBottomNavigationItem> bottomNavigationItems = new ArrayList<>();
 	private boolean useMenuResource = true;
 	private int[] tabColors;
+	private Handler handler = new Handler();
 
 	// UI
 	private AHBottomNavigationViewPager viewPager;
@@ -37,15 +41,28 @@ public class DemoActivity extends AppCompatActivity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
+		boolean enabledTranslucentNavigation = getSharedPreferences("shared", Context.MODE_PRIVATE)
+				.getBoolean("translucentNavigation", false);
+		setTheme(enabledTranslucentNavigation ? R.style.AppTheme_TranslucentNavigation : R.style.AppTheme);
 		setContentView(R.layout.activity_home);
 		initUI();
 	}
 
+	@Override
+	protected void onDestroy() {
+		super.onDestroy();
+		handler.removeCallbacksAndMessages(null);
+	}
+	
 	/**
 	 * Init UI
 	 */
 	private void initUI() {
-
+		
+		if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
+			AppCompatDelegate.setCompatVectorFromResourcesEnabled(true);
+		}
+		
 		bottomNavigation = (AHBottomNavigation) findViewById(R.id.bottom_navigation);
 		viewPager = (AHBottomNavigationViewPager) findViewById(R.id.view_pager);
 		floatingActionButton = (FloatingActionButton) findViewById(R.id.floating_action_button);
@@ -66,6 +83,9 @@ public class DemoActivity extends AppCompatActivity {
 			bottomNavigation.addItems(bottomNavigationItems);
 		}
 
+		bottomNavigation.manageFloatingActionButtonBehavior(floatingActionButton);
+		bottomNavigation.setTranslucentNavigationEnabled(true);
+
 		bottomNavigation.setOnTabSelectedListener(new AHBottomNavigation.OnTabSelectedListener() {
 			@Override
 			public boolean onTabSelected(int position, boolean wasSelected) {
@@ -84,6 +104,11 @@ public class DemoActivity extends AppCompatActivity {
 				}
 
 				viewPager.setCurrentItem(position, false);
+				
+				if (currentFragment == null) {
+					return true;
+				}
+				
 				currentFragment = adapter.getCurrentFragment();
 				currentFragment.willBeDisplayed();
 
@@ -161,12 +186,14 @@ public class DemoActivity extends AppCompatActivity {
 				return true;
 			}
 		});
-
+		
+		/*
 		bottomNavigation.setOnNavigationPositionListener(new AHBottomNavigation.OnNavigationPositionListener() {
 			@Override public void onPositionChange(int y) {
 				Log.d("DemoActivity", "BottomNavigation Position: " + y);
 			}
 		});
+		*/
 
 		viewPager.setOffscreenPageLimit(4);
 		adapter = new DemoViewPagerAdapter(getSupportFragmentManager());
@@ -174,12 +201,11 @@ public class DemoActivity extends AppCompatActivity {
 
 		currentFragment = adapter.getCurrentFragment();
 
-		final Handler handler = new Handler();
 		handler.postDelayed(new Runnable() {
 			@Override
 			public void run() {
 				// Setting custom colors for notification
-				Notification notification = Notification.newBuilder()
+				AHNotification notification = new AHNotification.Builder()
 						.setText(":)")
 						.setBackgroundColor(ContextCompat.getColor(DemoActivity.this, R.color.color_notification_back))
 						.setTextColor(ContextCompat.getColor(DemoActivity.this, R.color.color_notification_text))
@@ -187,9 +213,10 @@ public class DemoActivity extends AppCompatActivity {
 				bottomNavigation.setNotification(notification, 1);
 				Snackbar.make(bottomNavigation, "Snackbar with bottom navigation",
 						Snackbar.LENGTH_SHORT).show();
+
 			}
 		}, 3000);
-
+		
 		//bottomNavigation.setDefaultBackgroundResource(R.drawable.bottom_navigation_background);
 	}
 
@@ -257,6 +284,22 @@ public class DemoActivity extends AppCompatActivity {
 	 */
 	public void updateSelectedBackgroundVisibility(boolean isVisible) {
 		bottomNavigation.setSelectedBackgroundVisible(isVisible);
+	}
+
+	/**
+	 * Show or hide selected item background
+	 */
+	public void setForceTitleHide(boolean forceTitleHide) {
+		AHBottomNavigation.TitleState state = forceTitleHide ? AHBottomNavigation.TitleState.ALWAYS_HIDE : AHBottomNavigation.TitleState.ALWAYS_SHOW;
+		bottomNavigation.setTitleState(state);
+	}
+
+	/**
+	 * Reload activity
+	 */
+	public void reload() {
+		startActivity(new Intent(this, DemoActivity.class));
+		finish();
 	}
 
 	/**